chore: removed warning masks, perf: update TodoJson::{id: u16 => u32

+++ src/models.rs
@@ -1,9 +1,9 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Clone)]
-#[allow(dead_code)]
+
 pub struct TodoJson {
-    pub id: u16,
+    pub id: u32,
     pub title: String,
     pub completed: bool,
 }
